- Changed model directory - Moved all database related files to models.database
