+++ app/src/main/java/com/example/androidDeviceDetails/adapters/NetWorkUsageListAdapter.kt
@@ -6,8 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import com.example.androidDeviceDetails.R
-import com.example.androidDeviceDetails.models.networkUsageModels.AppNetworkUsageRaw
-import com.example.androidDeviceDetails.models.networkUsageModels.NetworkUsageItemViewHolder
+import com.example.androidDeviceDetails.models.database.AppNetworkUsageRaw
+import com.example.androidDeviceDetails.models.networkUsage.NetworkUsageItemViewHolder
 import com.example.androidDeviceDetails.utils.Utils
 
 class NetWorkUsageListAdapter(
