+++ app/src/main/java/com/example/androidDeviceDetails/collectors/AppEventCollector.kt
@@ -5,8 +5,8 @@ import android.app.usage.UsageStatsManager
 import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import com.example.androidDeviceDetails.base.BaseCollector
-import com.example.androidDeviceDetails.models.RoomDB
-import com.example.androidDeviceDetails.models.batteryModels.AppEventRaw
+import com.example.androidDeviceDetails.models.database.AppEventRaw
+import com.example.androidDeviceDetails.models.database.RoomDB
 import com.example.androidDeviceDetails.utils.Utils
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
