+++ app/src/main/java/com/example/androidDeviceDetails/collectors/BatteryCollector.kt
@@ -8,8 +8,8 @@ import android.os.BatteryManager
 import com.example.androidDeviceDetails.DeviceDetailsApplication
 import com.example.androidDeviceDetails.base.BaseCollector
 import com.example.androidDeviceDetails.collectors.BatteryCollector.BatteryReceiver
-import com.example.androidDeviceDetails.models.RoomDB
-import com.example.androidDeviceDetails.models.batteryModels.BatteryRaw
+import com.example.androidDeviceDetails.models.database.BatteryRaw
+import com.example.androidDeviceDetails.models.database.RoomDB
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
