+++ app/src/main/java/com/example/androidDeviceDetails/cooker/NetworkUsageCooker.kt
@@ -2,9 +2,9 @@ package com.example.androidDeviceDetails.cooker
 
 import com.example.androidDeviceDetails.base.BaseCooker
 import com.example.androidDeviceDetails.interfaces.ICookingDone
-import com.example.androidDeviceDetails.models.RoomDB
 import com.example.androidDeviceDetails.models.TimePeriod
-import com.example.androidDeviceDetails.models.networkUsageModels.AppNetworkUsageRaw
+import com.example.androidDeviceDetails.models.database.AppNetworkUsageRaw
+import com.example.androidDeviceDetails.models.database.RoomDB
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import java.util.*
