+++ app/src/main/java/com/example/androidDeviceDetails/interfaces/OnItemClickListener.kt
@@ -1,6 +1,6 @@
 package com.example.androidDeviceDetails.interfaces
 
-import com.example.androidDeviceDetails.models.locationModels.LocationDisplayModel
+import com.example.androidDeviceDetails.models.location.LocationDisplayModel
 
 interface OnItemClickListener {
     fun onItemClicked(clickedItem: LocationDisplayModel)
