+++ app/src/main/java/com/example/androidDeviceDetails/models/location/LocationDisplayModel.kt
@@ -1,4 +1,4 @@
-package com.example.androidDeviceDetails.models.locationModels
+package com.example.androidDeviceDetails.models.location
 
 data class LocationDisplayModel(
     val geoHash: String,
