+++ app/src/main/java/com/example/androidDeviceDetails/utils/AppInfoCollectionHelper.kt
@@ -1,8 +1,9 @@
 package com.example.androidDeviceDetails.utils
 
 import android.content.Context
-import com.example.androidDeviceDetails.models.RoomDB
-import com.example.androidDeviceDetails.models.appInfoModels.*
+import com.example.androidDeviceDetails.models.appInfo.AppDetails
+import com.example.androidDeviceDetails.models.appInfo.EventType
+import com.example.androidDeviceDetails.models.database.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
