+++ app/src/main/java/com/example/androidDeviceDetails/viewModel/SignalViewModel.kt
@@ -4,13 +4,13 @@ import android.annotation.SuppressLint
 import android.content.Context
 import androidx.core.view.isVisible
 import com.example.androidDeviceDetails.R
-import com.example.androidDeviceDetails.ui.SignalActivity
 import com.example.androidDeviceDetails.adapters.SignalListAdapter
 import com.example.androidDeviceDetails.base.BaseViewModel
 import com.example.androidDeviceDetails.cooker.SignalCooker
 import com.example.androidDeviceDetails.databinding.ActivitySignalBinding
-import com.example.androidDeviceDetails.models.RoomDB
-import com.example.androidDeviceDetails.models.signalModels.SignalRaw
+import com.example.androidDeviceDetails.models.database.RoomDB
+import com.example.androidDeviceDetails.models.signal.SignalRaw
+import com.example.androidDeviceDetails.ui.SignalActivity
 import com.example.androidDeviceDetails.utils.Signal
 
 /**
